backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dashscope>=1.22.0",
    "fastapi>=0.129.0",
    "jinja2>=3.1.6",
    "minio>=7.2.20",
    "numpy>=2.4.0",
    "openai>=2.14.0",
    "passlib[argon2,bcrypt]>=1.7.4",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pymilvus>=2.5.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "sqlmodel>=0.0.33",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.15.1",
]
